refactor: プレゼンテーション層内の階層整理
