+++ lib/utils/url_launcher_service.dart
@@ -2,7 +2,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import '../domain/app_exception.dart';
-import '../presentation/components/loading.dart';
+import '../presentation/common/components/loading.dart';
 
 /// URL起動サービスプロバイダー
 final urlLauncherServiceProvider = Provider(
